added dplyr compatibility comments to colpair_map()
dplyr v1.0.0-v1.0.3 had a different implementation for across() requiring some arguments to be masked with double curly brackets ({{) when I used it to implement colpair_map(). That is no longer the case for dplyr v1.0.4 and above, but I've kept the curly brackets there for compatibility with the earlier versions. They continue to work as before. I've now added code comments to explain why they are there in case anyone comes across this in the future.