no longer testing that `.data` is a valid column name for colpair_map().
The idea behind this unit test is to ensure user column names get masked. So that it doesn't matter what users name their columns. This is achievable for almost all column names except, since dplyr v1.0.4, `.data`. This is fine because `.data` is vanishingly unlikely to be used as a column name, hence why it's the first argument to colpair_map(). This also mirrors dplyr::summarise() and other dplyr functions, where you also cannot create a new column called `.data`. I'm quite suprised this ever worked in the first place for colpair_map(). So I've removed `.data` from the list of fields that we test against and added a code comment to specify that this is not a mistake.